internal/releasetargets: remove duplicate linux-arm target
The code that added missing secondary ports didn't know about the v6l
suffix. Teach it.

Also change the testing gcsfs implementation to reject overwrites so the
problem would have been caught in tests.

For golang/go#60921

Reviewed-on: https://go-review.googlesource.com/c/build/+/504820
Run-TryBot: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>

+++ internal/gcsfs/osfs.go
@@ -5,6 +5,8 @@
 package gcsfs
 
 import (
+	"errors"
+	"fmt"
 	"io/fs"
 	"os"
 	"path"
@@ -72,6 +74,9 @@ func (dir dirFS) Create(name string) (WriterFile, error) {
 		return nil, err
 	}
 	finalize := func() error {
+		if _, err := os.Stat(fullName); !errors.Is(err, fs.ErrNotExist) {
+			return fmt.Errorf("file exists and cannot be overwritten: %v", name)
+		}
 		return os.Rename(temp.Name(), fullName)
 	}
 	return &atomicWriteFile{temp, finalize}, nil
